+++ cms/src/main/java/dia/kids/cms/model/CommentDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CommentDto {
-    Integer articleId;
-    Integer userId;
+    Long articleId;
+    Long userId;
     String text;
 }
