+++ cms/src/main/java/dia/kids/cms/repository/ArticleRepository.java
@@ -1,10 +1,7 @@
 package dia.kids.cms.repository;
 
 import dia.kids.cms.model.Article;
-import dia.kids.cms.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface ArticleRepository extends JpaRepository<Article, Integer> {
+public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
