+++ cms/src/main/java/dia/kids/cms/repository/CommentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-    List<Comment> findAllByArticleId(Integer articleId);
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    List<Comment> findAllByArticleId(Long articleId);
 }
